move local storage handling to service
